requirements.txt
loguru==0.6.0
matplotlib==3.6.2
pandas==1.4.4
psycopg2-binary==2.9.3
seaborn==0.12.2
scikit-learn==1.1.2